Move RequestInterceptor to http.log
