+++ app/src/main/java/me/jessyan/mvparms/demo/app/GlobalHttpHandlerImpl.java
@@ -20,7 +20,7 @@ import android.text.TextUtils;
 
 import com.google.gson.reflect.TypeToken;
 import com.jess.arms.http.GlobalHttpHandler;
-import com.jess.arms.http.RequestInterceptor;
+import com.jess.arms.http.log.RequestInterceptor;
 import com.jess.arms.utils.ArmsUtils;
 
 import java.util.List;
