+++ arms/src/main/java/com/jess/arms/di/module/ClientModule.java
@@ -21,7 +21,7 @@ import android.support.annotation.Nullable;
 
 import com.google.gson.Gson;
 import com.jess.arms.http.GlobalHttpHandler;
-import com.jess.arms.http.RequestInterceptor;
+import com.jess.arms.http.log.RequestInterceptor;
 import com.jess.arms.utils.DataHelper;
 
 import java.io.File;
